pyproject.toml
[project]
name = "alma-agent"
version = "0.1.0"
description = "Alma Agent - Copiloto Pentester Inteligente"
authors = [
    {name = "Tu Nombre", email = "tu@email.com"},
]

dependencies = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "psycopg2-binary>=2.9.0",
    "langchain>=0.1.0",
    "langchain-community>=0.0.10",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "sqlalchemy>=2.0.0",
]

[project.scripts]
alma = "alma.__main__:app"

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"